% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convi.R
\name{convi}
\alias{convi}
\title{conversion interes mensual a interes anual}
\usage{
convi(im)
}
\arguments{
\item{im}{tasa de interes mensual. 0 < im < 1}
}
\value{
tasa de interes anual
}
\description{
conversion interes mensual a interes anual
}
\examples{
 library(Actuaria)
 convi(0.002709263)
}
