% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kqx.R
\name{kqx}
\alias{kqx}
\title{Esta funcion calcula kqx}
\usage{
kqx(x, k, sex, base)
}
\arguments{
\item{x}{Edad Inicial segun tabla TCM-89-03}

\item{k}{Años de Sobrevivencia segun tabla TCM-89,03}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
Función kqx
}
\description{
Esta funcion calcula kqx
}
\examples{
library(Actuaria)
x <- tcm8903
kqx(23,5,"M",x)
}
