% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensual.R
\name{mensual}
\alias{mensual}
\title{prima de un seguro para meses (interpolacion Lineal o hiperbolica)}
\usage{
mensual(SA, x, n, m, i, sex, interpolar = "lin", base)
}
\arguments{
\item{SA}{Suma asegurada}

\item{x}{Edad Inicial segùn tabla TCM-89-03 (En meses)}

\item{n}{Meses Temporalidad}

\item{m}{Meses Diferimiento}

\item{i}{tasa de interes (Efectiva Anual). 0 < i < 1. Si tiene la tasa de interes mensual utilize la función interes para saber su equivalente anual}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{interpolar}{Metodo de interpolacion "lin" para lineal o "hip" para hiperbolica}

\item{base}{Base de datos que contiene lx}
}
\value{
prima de un seguro para meses
}
\description{
prima de un seguro para meses (interpolacion Lineal o hiperbolica)
}
\examples{
 library(Actuaria)
 base <-  tcm8903

 Interpolación Lineal
 mensual(SA = 150000000, x = 360,n = 0,m = 0,i = 0.033,sex = "M",interpolar = "lin",base )

 Interpolacion hiperbolica
 mensual(SA = 150000000, x = 360,n = 0,m = 0,i = 0.033,sex = "M",interpolar = "hip",base )
}
