% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensualdot.R
\name{mensualdot}
\alias{mensualdot}
\title{prima de un seguro dotal para meses (interpolacion Lineal o hiperbolica)}
\usage{
mensualdot(SA, B, x, n, m, i, sex, interpolar = "lin", base)
}
\arguments{
\item{SA}{Suma asegurada}

\item{B}{Suma asegurada dotal}

\item{x}{Edad Inicial segun tabla TCM-89-03 (En meses)}

\item{n}{Meses Temporalidad}

\item{m}{Meses Diferimiento}

\item{i}{tasa de interes (Efectiva Anual). 0 < i < 1. Si tiene la tasa de interes mensual utilize la función convi para saber su equivalente anual}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{interpolar}{Metodo de interpolacion "lin" para lineal o "hip" para hiperbolica}

\item{base}{Base de datos que contiene lx}
}
\value{
prima de un seguro para meses
}
\description{
prima de un seguro dotal para meses (interpolacion Lineal o hiperbolica)
}
\examples{
 library(Actuaria)
 base <-  tcm8903

 Interpolación Lineal
 mensualdot(100000000,10000000,788,159,24,0.0345,"M", "lin", base)

 Interpolacion hiperbolica
 mensualdot(100000000,10000000,788,159,24,0.0345,"M", "hip", base)
}
