% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prima.R
\name{prima}
\alias{prima}
\title{Calcula la prima de un seguro por muerte}
\usage{
prima(SA, x, n, m, i, sex, base)
}
\arguments{
\item{SA}{Suma asegurada}

\item{x}{Edad Inicial segùn tabla TCM-89-03}

\item{n}{Años Temporalidad}

\item{m}{Años Diferimiento}

\item{i}{tasa de interes. 0 < i < 1}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
prima de un seguro por muerte
}
\description{
Calcula la prima de un seguro por muerte
}
\examples{
library(Actuaria)

#Seguro Anual
base <- tcm8903
prima(100000000,30,0,0,0.025,"M",base)

#Seguro Anual con temporalidad
prima(100000000,23,20,0,0.025,"F", base)

#Seguro  Anual con temporalidad y diferimiento
prima(100000000,55,15,8,0.03,"M",base)

}
