% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primadot.R
\name{primadot}
\alias{primadot}
\title{Calcula la prima de un seguro dotal}
\usage{
primadot(SA, B, x, n, m, i, sex, base)
}
\arguments{
\item{SA}{Suma asegurada por muerte}

\item{B}{Suma asegurada por sobrevivencia}

\item{x}{Edad Inicial segùn tabla TCM-89-03}

\item{n}{Años Temporalidad}

\item{m}{Años Diferimiento}

\item{i}{tasa de interes. 0 < i < 1}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
prima de un seguro por muerte
}
\description{
Calcula la prima de un seguro dotal
}
\examples{
library(Actuaria)

base <- tcm8903

primadot(100000000,12000000,55,15,8,0.03,"M",base)

}
