% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{tcm8903}
\alias{tcm8903}
\title{Base de Datos tcm-89-03}
\format{A data frame with 81 rows and 3 variables:
\itemize{
  \item edad: Edad
  \item lx: Sobrevivencia Hombre
  \item ly: Sobrevivencia Mujer
}}
\usage{
tcm8903
}
\description{
Base de Datos entragada por Deisy Camargo
}
\keyword{datasets}
