% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpa.R
\name{vpa}
\alias{vpa}
\title{Valor presente Actuarial}
\usage{
vpa(R, x, n, i, sex, base)
}
\arguments{
\item{R}{Suma asegurada}

\item{x}{Edad Inicial}

\item{n}{Años Temporalidad}

\item{i}{tasa de interes. 0 < i < 1}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
valor presente Actuarial
}
\description{
Valor presente Actuarial
}
\examples{
library(Actuaria)

#Valor presente actuarial Anual
base <- tcm8903
vpa(100000000,30,0,0.025,"M",base)

#Valor presente actuarial Anual con temporalidad
vpa(100000000,23,20,0.025,"F", base)


}
