% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpamensual.R
\name{vpamensual}
\alias{vpamensual}
\title{Valor Presente Actuarial para meses (interpolacion Lineal o hiperbolica)}
\usage{
vpamensual(R, x, n, i, sex, interpolar = "lin", base)
}
\arguments{
\item{R}{Suma asegurada}

\item{x}{Edad Inicial segùn tabla TCM-89-03 (En meses)}

\item{n}{Meses Temporalidad}

\item{i}{tasa de interes (Efectiva Anual). 0 < i < 1. Si tiene la tasa de interes mensual utilize la función interes para saber su equivalente anual}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{interpolar}{Metodo de interpolacion "lin" para lineal o "hip" para hiperbolica}

\item{base}{Base de datos que contiene lx}
}
\value{
prima de un seguro para meses
}
\description{
Valor Presente Actuarial para meses (interpolacion Lineal o hiperbolica)
}
\examples{
 library(Actuaria)
 base <-  tcm8903

 Valor Presente Actuarial Interpolación Lineal
 mensual(R = 150000000, x = 360,n = 0,i = 0.033,sex = "M",interpolar = "lin",base )

 Valor Presente Actuarial Interpolacion hiperbolica
 mensual(R = 150000000, x = 360,n = 0,i = 0.033,sex = "M",interpolar = "hip",base )
}
