% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kPx.R
\name{kPx}
\alias{kPx}
\title{calcula kPx}
\usage{
kPx(x, k, sex, base)
}
\arguments{
\item{x}{Edad Inicial segùn tabla TCM-89,03}

\item{k}{Años de Sobrevivencia segùn tabla TCM-89,03}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
kPx
}
\description{
calcula kPx
}
\examples{
library(Actuaria)
x <- tcm8903
kPx(23,5,"F",x)
}
