% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prima.R
\name{prima}
\alias{prima}
\title{Esta funcion calcula la prima de un seguro por muerte}
\usage{
prima(SA, x, n, m, i, sex, base)
}
\arguments{
\item{SA}{Suma asegurada}

\item{x}{Edad Inicial segun tabla TCM-89-03}

\item{n}{Años Temporalidad}

\item{m}{Años Diferimiento}

\item{i}{tasa de interes. 0 < i < 1}

\item{sex}{Sexo "M" para masculino, "F" para Femenino}

\item{base}{Base de datos que contiene lx}
}
\value{
Función kqx
}
\description{
Esta funcion calcula la prima de un seguro por muerte
}
\examples{
library(Actuaria)
x <- tcm8903
kqx(23,5,"M",x)
}
